refactor: prop types 적용, 컴포넌트 폴더 이름 변경
