+++ package.json
@@ -6,6 +6,7 @@
   "author": "ddongule <mnk918p@gmail.com>",
   "license": "MIT",
   "dependencies": {
+    "proptypes": "^1.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-lottie": "^1.2.3"
