+++ src/components/MoneyInput/style.scss
