+++ src/components/TimeLeft/style.scss
@@ -10,6 +10,7 @@
   font-family: 'Noto Serif KR', serif;
   color: white;
   z-index: 11;
+  white-space: nowrap;
 }
 
 .belt-up {
