+++ src/components/WinningNumber/style.scss
