+++ src/components/receipt/LotteryBall/index.js
