+++ src/components/receipt/LotteryBall/style.scss
