+++ src/components/receipt/PurchaseNumberItem/style.scss
