+++ src/components/receipt/PurchaseNumberList/style.scss
